// src/app/data/models/arbol-qa.ts
export class ArbolQa {
    id?: number;
    sucursal: string;
    zona: string;
    pagoNoRegistrado: string;
    domicilioConAdeudo: string;
    cargosImprocedentes: string;
    folioNoIngresado: string;
    aclaracionVenta: string;
    contratoDuplicado: string;
    imposibilidadTecnica: string;
    danosPropiedad: string;
    quejaRepetitival: string;
    mantenimientoPendiente: string;
    equipoNoRegistrado: string;
    trabajoPro: string;
    cambioCiudad: string;
    validarTrabajoAtendido: string;
    auditoria: string;
    malServicoCis: string;
    malServicioCambaceo: string;
    malServicioTecnico: string;
    aplicarPagosEquivocados: string;
    quejaPorRetMegared: string;
    quejaPorRetCable: string;
    portabilidadMal: string;
}